+++ telnetlib3/tests/test_client_unit.py
@@ -6,7 +6,7 @@ import pytest
 
 # local
 from telnetlib3 import client as cl
-from telnetlib3.tests.accessories import (  # noqa: F401  # pylint: disable=unused-import
+from telnetlib3.tests.accessories import (  # noqa: F401
     bind_host,
     create_server,
 )