+++ telnetlib3/tests/test_core.py
@@ -15,7 +15,7 @@ import pexpect
 # local
 import telnetlib3
 from telnetlib3.telopt import DO, SB, IAC, SGA, NAWS, WILL, WONT, TTYPE, BINARY, CHARSET
-from telnetlib3.tests.accessories import (  # pylint: disable=unused-import
+from telnetlib3.tests.accessories import (
     bind_host,
     create_server,
     asyncio_server,
@@ -499,7 +499,7 @@ async def test_telnet_client_cmdline_stdin_pipe(bind_host, unused_tcp_port):
             logfile_output = f.read().splitlines()
         assert stdout == (
             b"Escape character is '^]'.\n"
-            b"Press Return to continue:\r\ngoodbye.\r\n"
+            b"Press Return to continue:\ngoodbye.\n"
             b"\x1b[m\nConnection closed by foreign host.\n"
         )
         assert len(logfile_output) in (2, 3), logfile