+++ telnetlib3/tests/test_fingerprinting.py
@@ -22,7 +22,7 @@ else:
     server_pty_shell = None  # type: ignore[assignment]
 
 # local
-from telnetlib3.tests.accessories import (  # noqa: F401  # pylint: disable=unused-import
+from telnetlib3.tests.accessories import (  # noqa: F401
     bind_host,
     create_server,
     open_connection,