+++ telnetlib3/tests/test_naws.py
@@ -17,7 +17,7 @@ import pexpect
 # local
 import telnetlib3
 from telnetlib3.telopt import SB, SE, IAC, NAWS, WILL
-from telnetlib3.tests.accessories import (  # pylint: disable=unused-import; pylint: disable=unused-import,
+from telnetlib3.tests.accessories import (
     bind_host,
     create_server,
     open_connection,