+++ telnetlib3/tests/test_reader.py
@@ -8,7 +8,7 @@ import pytest
 
 # local
 import telnetlib3
-from telnetlib3.tests.accessories import (  # pylint: disable=unused-import; pylint: disable=unused-import,
+from telnetlib3.tests.accessories import (
     bind_host,
     create_server,
     open_connection,