+++ telnetlib3/tests/test_shell.py
@@ -7,7 +7,7 @@ import logging
 # local
 from telnetlib3 import accessories, telnet_server_shell
 from telnetlib3.telopt import DO, IAC, SGA, ECHO, WILL, WONT, TTYPE, BINARY
-from telnetlib3.tests.accessories import (  # pylint: disable=unused-import
+from telnetlib3.tests.accessories import (
     bind_host,
     create_server,
     asyncio_server,